pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "boss-tnc-style"
version = "0.1.0"
description = "BOSS-TNC paper system organized in experiment-style layout"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "networkx>=3.0",
  "opt_einsum>=3.3",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "PyYAML>=6.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
